Add Layout, Text, Icons and local databases
Why is it necessary?
This feature strengthen the Front-End architeture of the system

How does the change address the issuer?
Components.js -> Created an IconComponent with dynamic values to conentrate Icon logic in one class
HomeItemLayout -> The layout of each card on the main menu
assets/data -> Create Json files that mimic the data on the AWS cloud
Styles.js -> Create new dynamic values for the card coponent so it can easily be reused

What side effect does this change have ?
Icons do not display correctly, they may have to be fix in the future